requirements.txt
asgiref==3.4.1
Django==4.0
django-js-asset==1.2.2
django-mptt==0.13.4
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
jsonfield==3.1.0
phonenumbers==8.12.39
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
